+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend/shared_helpers/shared_helpers_test.go
@@ -11,11 +11,11 @@ func TestGetServicePortsFromPortSpecs(t *testing.T) {
 	portId2 := "udp"
 	portId3 := "sctp"
 
-	portSpec1, err := port_spec.NewPortSpec(100, port_spec.PortProtocol_TCP)
+	portSpec1, err := port_spec.NewPortSpec(100, port_spec.PortProtocol_TCP, "")
 	require.NoError(t, err)
-	portSpec2, err := port_spec.NewPortSpec(200, port_spec.PortProtocol_UDP)
+	portSpec2, err := port_spec.NewPortSpec(200, port_spec.PortProtocol_UDP, "")
 	require.NoError(t, err)
-	portSpec3, err := port_spec.NewPortSpec(300, port_spec.PortProtocol_SCTP)
+	portSpec3, err := port_spec.NewPortSpec(300, port_spec.PortProtocol_SCTP, "")
 	require.NoError(t, err)
 
 	_, err = GetKubernetesServicePortsFromPrivatePortSpecs(map[string]*port_spec.PortSpec{
@@ -31,11 +31,11 @@ func TestGetContainerPortsFromPortSpecs(t *testing.T) {
 	portId2 := "udp"
 	portId3 := "sctp"
 
-	portSpec1, err := port_spec.NewPortSpec(100, port_spec.PortProtocol_TCP)
+	portSpec1, err := port_spec.NewPortSpec(100, port_spec.PortProtocol_TCP, "")
 	require.NoError(t, err)
-	portSpec2, err := port_spec.NewPortSpec(200, port_spec.PortProtocol_UDP)
+	portSpec2, err := port_spec.NewPortSpec(200, port_spec.PortProtocol_UDP, "")
 	require.NoError(t, err)
-	portSpec3, err := port_spec.NewPortSpec(300, port_spec.PortProtocol_SCTP)
+	portSpec3, err := port_spec.NewPortSpec(300, port_spec.PortProtocol_SCTP, "")
 	require.NoError(t, err)
 
 	_, err = GetKubernetesContainerPortsFromPrivatePortSpecs(map[string]*port_spec.PortSpec{