+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/add_service/add_service_instruction.go
@@ -198,7 +198,7 @@ func (instruction *AddServiceInstruction) ValidateAndUpdateEnvironment(environme
 		}
 	}
 	environment.AddServiceId(instruction.serviceId)
-	environment.AppendRequiredDockerImage(instruction.serviceConfig.ContainerImageName)
+	environment.AppendRequiredContainerImage(instruction.serviceConfig.ContainerImageName)
 	return nil
 }